# Add doc-value aggregator for variance aggregation.