## Summary of the changes / Why this improves CrateDB
Before:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global variance-long
   Statement:
     select variance(duration) from uservisits
   Concurrency: 1
   Iterations: 500
100%|███████████████████████████████████████████████████████████| 500/500 [00:30<00:00, 16.62 requests/s]
Runtime (in ms):
    mean:    58.263 ± 3.359
    min/max: 37.833 → 730.229
Percentile:
    50:   48.920 ± 38.327 (stdev)
    95:   104.321
    99.9: 730.229

## Running Query:
   Name: global variance-double
   Statement:
     select variance("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 500
100%|███████████████████████████████████████████████████████████| 500/500 [00:33<00:00, 14.98 requests/s]
Runtime (in ms):
    mean:    64.909 ± 2.987
    min/max: 46.892 → 540.260
Percentile:
    50:   58.078 ± 34.077 (stdev)
    95:   89.473
    99.9: 540.260
```

After:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global variance-long
   Statement:
     select variance(duration) from uservisits
   Concurrency: 1
   Iterations: 500
100%|███████████████████████████████████████████████████████████| 500/500 [00:11<00:00, 43.70 requests/s]
Runtime (in ms):
    mean:    21.141 ± 0.333
    min/max: 16.002 → 42.931
Percentile:
    50:   20.051 ± 3.797 (stdev)
    95:   29.713
    99.9: 42.931

## Running Query:
   Name: global variance-double
   Statement:
     select variance("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 500
100%|███████████████████████████████████████████████████████████| 500/500 [00:12<00:00, 38.94 requests/s]
Runtime (in ms):
    mean:    24.418 ± 2.144
    min/max: 17.440 → 329.129
Percentile:
    50:   21.321 ± 24.460 (stdev)
    95:   30.445
    99.9: 329.129
```

## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [ ] Touched code is covered by tests
 - [ ] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [ ] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)